fix validation and resp
